+++ examples/websocket/Cargo.toml
@@ -7,8 +7,9 @@ edition = "2021"
 
 [dependencies]
 async-trait = "0.1.68"
-silent = { path = "../../silent" }
+silent = { path = "../../silent", features = ["ws"] }
 tokio = { version = "1", features = ["full"] }
 tokio-tungstenite = "0.18.0"
 futures-util = { version = "0.3.28" }
 backtrace = "0.3"
+headers = "0.3.8"
