+++ silent/src/ws/mod.rs
@@ -1,0 +1,5 @@
+mod handler_wrapper_websocket;
+mod message;
+mod websocket;
+
+pub use handler_wrapper_websocket::HandlerWrapperWebSocket;
